requirements.txt
dj-database-url==0.4.1
Django==1.11.1
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
django-flat-responsive==1.4.1
icalendar==3.11.5
requests==2.18.1
python-dateutil==2.6.1